excel-mock-interviewer-advanced/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1
pandas==2.2.2
SQLAlchemy==2.0.32
openai==1.37.0
httpx==0.27.2
